Rapids new Feature: G0 codes have a fixed feedrate
G0 was being executed in my FABtotum at the same feedrate as G1. This was very slow for a rapid.

This version explicitly sets a fixed feedrate for rapids. In current version 2000 mm/min.

This number is initialised in the constructor and you may edit the file and set another value of your choice.