Bug fix: Handling of Preps (G90...) Bug fix: Safety travel for profile
Handling of Preps was done during a GX command (G0 G90 X90), now it correctly handles the G90 separately.
This is specially relevant for multiple sketches profiling (for example insulation milling of a PCB)

Heekscad does not seem to incorporate an initial tool travel Z for profiles, which causes the milling to scratch the work-part (e.g. PCB) until the initial point of the profile.
This only happens for the first profile (afterwards travel Z is correctly set). I solved this by hardcoding a G0 Z40 before the first rapid. I am not proud of the hardcoding of
40 mm, but I failed to see how to read the value set in the software for Z travel. At least it won't scratch the PCB now.